Split api routes by applications
