+++ meetupselector/api/routes/__init__.py
@@ -1,0 +1,3 @@
+from .proposals import router as proposals_router  # noqa: F401
+from .talks import router as talks_router  # noqa: F401
+from .users import router as users_router  # noqa: F401
