+++ meetupselector/api/routes/users.py
@@ -1,0 +1,6 @@
+from ninja import Router
+
+router = Router()
+
+
+# Add endpoints here. Example:  @router.get("/") ...
