+++ meetupselector/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
-from meetupselector.api.routes import api
+from meetupselector.api.api import api
 
 urlpatterns = [
     path("admin/", admin.site.urls),
