+++ tests/api/routes/test_proposals.py
@@ -1,0 +1,1 @@
+"""Place here tests for proposals related endpoints"""
