+++ tests/api/routes/test_talks.py
@@ -1,0 +1,1 @@
+"""Place here tests for talks related endpoints"""
