+++ tests/api/routes/test_users.py
@@ -1,0 +1,1 @@
+"""Place here tests for users related endpoints"""
